Add files via upload
NFC combined with PIR. The module wakes up when PIR sensor detects movement. The NFC reader tries to find an RFID tag for 5 seconds. If no tag is detected within 5 seconds the LED flashes red and the module goes back to sleep. When a tag is detected the LED flashes green and goes back to sleep.